requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-docx>=0.8.11
openpyxl>=3.1.0
openai>=1.0.0
plotly>=5.14.0
wordcloud>=1.9.0
python-dotenv>=1.0.0
networkx>=3.0
scikit-learn>=1.3.0
pymongo>=4.3.0
matplotlib>=3.7.0
nltk>=3.8.0
textblob>=0.17.0
spacy>=3.5.0
transformers>=4.30.0
yake>=0.4.8
torch>=2.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
beautifulsoup4>=4.12.0
chardet>=5.0.0
markdown>=3.4.0
lxml>=4.9.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6
